+++ scripts/autogen/templates/message.h
@@ -14,6 +14,7 @@
 #include "roq/span.h"
 #include "roq/string.h"
 
+#include "roq/event.h"
 #include "roq/message_info.h"
 
 {% include 'includes' %}
@@ -21,52 +22,26 @@
 {% include 'namespace_begin' %}
 
 struct ROQ_PUBLIC {{ name }} final {
+  {{ name }}() = default;
+  {{ name }}({{name }}&&) = default;
+  {{ name }}(const {{name }}&) = delete;
+
 {% include 'variables' %}
 };
-
-struct ROQ_PUBLIC {{ name }}Event final {
-  const MessageInfo& message_info;
-  const {{ name }}& {{ filename }};
-};
-
-template <>
-inline {{ name }}Event create_event(
-    const MessageInfo& message_info,
-    const {{ name }}Event& event) {
-  return decltype(event) {
-    .message_info = message_info,
-    .{{ filename }} = event.{{ filename }},
-  };
-}
-
-namespace detail {
-template <>
-struct event_value_helper<{{ name }}Event> final {
-  using type = {{ name }};
-  explicit event_value_helper(const {{ name }}Event& event)
-      : _event(event) {
-  }
-  operator const type&() const {
-    return _event.{{ filename }};
-  }
- private:
-  const {{ name }}Event& _event;
-};
-}  // namespace detail
 
 {% include 'namespace_end' %}
 
 {% include 'format' %}
 
 template <>
-struct fmt::formatter<{{ namespaces | join('::') }}::{{ name }}Event> {
+struct fmt::formatter<{{ namespaces | join('::') }}::Event<{{ namespaces | join('::') }}::{{ name }}> > {
   template <typename Context>
   constexpr auto parse(Context& context) {
     return context.begin();
   }
   template <typename Context>
   auto format(
-      const {{ namespaces | join('::') }}::{{ name }}Event& value,
+      const {{ namespaces | join('::') }}::Event<{{ namespaces | join('::') }}::{{ name }}>& event,
       Context& context) {
     return format_to(
         context.out(),
@@ -78,7 +53,7 @@ struct fmt::formatter<{{ namespaces | join('::') }}::{{ name }}Event> {
 {%- raw %}
         R"(}})",
 {% endraw %}
-        value.message_info,
-        value.{{ filename }});
+        event.message_info,
+        event.value);
   }
 };